Activityのパッケージ構成修正
