+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/film/FilmActivity.kt
@@ -1,18 +1,18 @@
-package com.sw.sw_api_kotlin_project.activity.people
+package com.sw.sw_api_kotlin_project.screen.film
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.sw.sw_api_kotlin_project.databinding.ActivityPeopleBinding
+import com.sw.sw_api_kotlin_project.databinding.ActivityFilmBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class PeopleActivity : AppCompatActivity() {
-    private var _binding: ActivityPeopleBinding? = null
+class FilmActivity : AppCompatActivity() {
+    private var _binding: ActivityFilmBinding? = null
     private val binding get() = checkNotNull(_binding)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        _binding = ActivityPeopleBinding.inflate(layoutInflater)
+        _binding = ActivityFilmBinding.inflate(layoutInflater)
         setContentView(binding.root)
     }
