+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/people/PeopleActivity.kt
@@ -1,18 +1,18 @@
-package com.sw.sw_api_kotlin_project.activity.film
+package com.sw.sw_api_kotlin_project.screen.people
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.sw.sw_api_kotlin_project.databinding.ActivityFilmBinding
+import com.sw.sw_api_kotlin_project.databinding.ActivityPeopleBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class FilmActivity : AppCompatActivity() {
-    private var _binding: ActivityFilmBinding? = null
+class PeopleActivity : AppCompatActivity() {
+    private var _binding: ActivityPeopleBinding? = null
     private val binding get() = checkNotNull(_binding)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        _binding = ActivityFilmBinding.inflate(layoutInflater)
+        _binding = ActivityPeopleBinding.inflate(layoutInflater)
         setContentView(binding.root)
     }
