+++ SW_API_Kotlin_Project/app/src/main/res/layout/activity_film.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.film.FilmActivity">
+        tools:context=".screen.film.FilmActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/nav_film_host_fragment"
