+++ SW_API_Kotlin_Project/app/src/main/res/layout/activity_film_details.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.film.FilmDetailsActivity">
+        tools:context=".screen.film.FilmDetailsActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/nav_film_details_host_fragment"
