+++ SW_API_Kotlin_Project/app/src/main/res/layout/activity_people.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.people.PeopleActivity">
+        tools:context=".screen.people.PeopleActivity">
 
 
         <androidx.fragment.app.FragmentContainerView
