+++ SW_API_Kotlin_Project/app/src/main/res/layout/activity_people_details.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.people.PeopleDetailsActivity">
+        tools:context=".screen.people.PeopleDetailsActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/nav_people_details_host_fragment"
