+++ SW_API_Kotlin_Project/app/src/main/res/layout/activity_planet.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.planet.PlanetActivity">
+        tools:context=".screen.planet.PlanetActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/nav_people_host_fragment"
